Revert "move becker port from vcc/devices/becker to vcc/devices/serial"
This reverts commit b22d63b74abda44e1b469f713ef5880e4824b413.
